Rename persistAnce -> persistEnce
