+++ config/slips.yaml
@@ -85,7 +85,7 @@ parameters:
    # By default False. Meaning we don't DELETE the DB by default.
    deletePrevdb : True
    # You can remember the data in all the previous runs of the DB if you put False.
-   # Redis will remember as long as the redis server is not down. The persistance is
+   # Redis will remember as long as the redis server is not down. The persistence is
    # on the memory, not disk.
    # deletePrevdb : False
